backend-api/requirements.txt
django
django-ninja
django-ninja-jwt[crypto]
pydantic[email]
rav